Core - API - Work towards initial rework of printouts

+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/service/mongo/MongoObjectService.java
@@ -225,7 +225,7 @@ public abstract class MongoObjectService<T extends MainObject, S extends SearchO
 		
 		return new SearchResult<>(
 			list,
-			this.count(oqmDbIdOrName, filter),
+			(int) this.count(oqmDbIdOrName, filter),
 			!filters.isEmpty(),
 			pagingOptions,
 			searchObject
