+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/service/mongo/TopLevelMongoService.java
@@ -190,7 +190,7 @@ public abstract class TopLevelMongoService<T extends MainObject, S extends Searc
 		
 		return new SearchResult<>(
 			list,
-			this.count(filter),
+			(int) this.count(filter),
 			!filters.isEmpty(),
 			pagingOptions,
 			searchObject
